Rename files: use hyphen instead of underscore
