+++ benchmarks/bench-qsort.cpp
@@ -1,0 +1,4 @@
+#include "bench-qsort.hpp"
+#include "bench-argsort.hpp"
+#include "bench-partial-qsort.hpp"
+#include "bench-qselect.hpp"
