--- benchmarks/bench_qsort.cpp
@@ -1,4 +1,0 @@
-#include "bench_qsort.hpp"
-#include "bench_argsort.hpp"
-#include "bench_partial_qsort.hpp"
-#include "bench_qselect.hpp"
