+++ tests/meson.build
@@ -3,8 +3,8 @@ libtests = []
 if cpp.has_argument('-march=skylake-avx512')
   libtests += static_library('tests_kv',
     files(
-      'test_keyvalue.cpp',
-      'test_argsort.cpp',
+      'test-keyvalue.cpp',
+      'test-argsort.cpp',
     ),
     dependencies: gtest_dep,
     include_directories : [src, utils],
@@ -14,7 +14,7 @@ endif
 
 if cpp.has_argument('-march=icelake-client')
   libtests += static_library('tests_qsort',
-    files('test_sort.cpp', ),
+    files('test-qsort.cpp', ),
     dependencies: gtest_dep,
     include_directories : [src, utils],
     cpp_args : ['-O3', '-march=icelake-client'],
@@ -23,7 +23,7 @@ endif
 
 if cancompilefp16
   libtests += static_library('tests_qsortfp16',
-    files('test_qsortfp16.cpp', ),
+    files('test-qsortfp16.cpp', ),
     dependencies: gtest_dep,
     include_directories : [src, utils],
     cpp_args : ['-O3', '-march=sapphirerapids'],
